+++ src/effects/specialFeats.js
@@ -80,7 +80,6 @@ import { songOfVictoryEffect } from "./feats/songOfVictory.js";
 import { twinklingConstellationsEffect } from "./feats/twinklingConstellations.js";
 import { armsOfTheAstralSelfEffect } from "./feats/armsOfTheAstralSelf.js";
 import { ghostWalkEffect } from "./feats/ghostWalk.js";
-import { darkOnesOwnLuckffect } from "./feats/darkOnesOwnLuck.js";
 import { foeSlayerEffect } from "./feats/foeSlayer.js";
 
 
@@ -153,11 +152,6 @@ async function midiFeatureEffects(ddb, character, document) {
     }
     case "Crossbow Expert": {
       document = crossbowExpertEffect(document);
-      break;
-    }
-    case "Dark One's Own Luck":
-    case "Dark Ones Own Luck": {
-      document = darkOnesOwnLuckffect(document);
       break;
     }
     case "Daunting Roar": {