+++ src/parser/enrichers/feature/sorcerer/RevelationInFlesh.js
@@ -37,7 +37,7 @@ export default class RevelationInFlesh extends DDBEnricherMixin {
             durationSeconds: 600,
           },
           changes: [
-            DDBEnricherMixin.generateUpgradeChange("2 * @attributes.movement.walk", 20, "system.attributes.movement.swim"),
+            this.movementChange("2 * @attributes.movement.walk", 20, "system.attributes.movement.swim"),
           ],
         },
         {
@@ -46,7 +46,7 @@ export default class RevelationInFlesh extends DDBEnricherMixin {
             durationSeconds: 600,
           },
           changes: [
-            DDBEnricherMixin.generateUpgradeChange("@attributes.movement.walk", 20, "system.attributes.movement.fly"),
+            this.movementChange("@attributes.movement.walk", 20, "system.attributes.movement.fly"),
             DDBEnricherMixin.generateCustomChange("true", 20, "system.attributes.movement.hover"),
           ],
         },