+++ src/parser/enrichers/trait/dragonborn/BreathWeapon2024.js
@@ -1,5 +1,5 @@
 /* eslint-disable class-methods-use-this */
-import DDBEnricherMixin from "../DDBEnricherMixin.js";
+import DDBEnricherMixin from "../../DDBEnricherMixin.js";
 
 export default class BreathWeapon2024 extends DDBEnricherMixin {